Use wg-quick to make a persistent lv wg interface
The wg-quick subsystem in the wireguard-tools package allows you to
define configuration files for wireguard interfaces.

Add functionality to make a wglv0 interface for use by labview, with a
sane ipv4 address and keypair. Add an initscript to start/stop the
interface. And exclude the interface from getting messed with by
ifplugd.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>