_firewall_config.py: work-out, public-out: allow basic ICMPv6 traffic
firewalld ships with a policy "allow-host-ipv6". Interestingly, it does *not*
specifically "enable IPv6". Instead, it merely accepts four ICMPv6 packets
through the firewall — neighbor advertisement/soliciation, router advertisement,
and redirect — and even then, ONLY for inbound packets. Nevertheless, they're
not particularly wrong: without neighbor advertisement/solicitation, IPv6 is
dead as a doorknob.

This is not going to be sufficient for our use cases, because we are setting up
outbound rules too. So we need to explicitly add neighbor
advertisement/solicitation in order for IPv6 to work. (We don't need to add any
inbound rules, because those in allow-host-ipv6 already match those.)

I also found that I needed to explicitly allow ICMPv6 echo-request and
echo-reply through, including for allow-host-ipv6.

Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>