remote-test-firewall.sh: add
This is a shell script for testing SNAC firewalld configuration between the
local machine and some remote SNAC system. TCP, UDP, ICMP, IPv4, and IPv6 are
tested in both directions. The eventual goal is to utilize this in an autotest
but its present use is to accelerate development-time testing.

Because this is currently a manual-only test (unlike the other tests under
`tests/integration/`), and should not be run by `make installcheck`, it is
located under `tests/system/` instead.

Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>