nilrt-snac entrypoint: defend against unsafe path
When calling the nilrt-snac wrapper script from a directory that
contains subdirectories called 'nilrt_snac', the python interpreter will
attempt to search those local paths for the nilrt_snac module and import
their contents. Instead, we should only load the nilrt_snac module from
its install locations.

Move the calculated install path to the front of the PYTHONPATH, so that
it is loaded first. And add `-P` to the python interpreter invocation,
which suppresses python's default behavior of adding CWD to the search
path.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>