gen_ula.py: add
It is useful to test an IPv6 firewall using global-scope addresses in addition
to link-scope. IPv6 mandates that we always get link-local addresses, but the
IPv6 equivalent of an IPv4 private address, a unique local address (ULA), does
not appear to be supported by basic Linux commands. So I have written my own
ULA generation utility here. Usage:

$ gen_ula.py --eui48 01:23:45:67:89:ab
fd35:bc81:941c:0:323:45ff:fe67:89ab/64

$ gen_ula.py --eui48 cc:dd:ee:ff:00:11 --network fd35:bc81:941c:0:323:45ff:fe67:89ab/64
fd35:bc81:941c:0:cedd:eeff:feff:11/64

Strictly speaking, we don't need ULAs if the DUT is on an isolated network; but
the simpler alternative is to hardcode the entire IPv6 address, which strikes me
as extremely poor form. Implementing the RFC by the book could prove useful for
both testing and system configuration in the future.

Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>